requirements.txt
django>=3.2
mysqlclient>=2.2.1
python-dotenv
djangorestframework
whitenoise
PyJWT
Pillow
channels
channels-redis
redis
daphne
